requirements.txt
streamlit>=1.32
python-dotenv>=1.0
pydantic>=2.6
requests>=2.31
huggingface_hub>=0.23.0